fix hashagg spill issue

Created-by: liuyu 00809192
Author-id: 53630
MR-id: 21196742
Commit-by: liuyu
Merged-by: ruanrunxue 00424934
E2E-issues: issue643;AR20240412519170
Description: + [ ] I've added/changed the documentation
+ [ ] I've added test for the code

See merge request: Kunpeng-Computing-DC/BigData/OmniOperatorJIT!774

+++ core/test/codegen/function_test.cpp
@@ -31,7 +31,7 @@ TEST(FunctionTest, ArenaAllocatorMalloc)
     auto execContext = new ExecutionContext();
     int64_t contextptr = reinterpret_cast<int64_t>(execContext);
     char *ptr;
-    for (int i = 1; i <= execContext->GetArena()->TotalBytes() / 256; ++i) {
+    for (uint64_t i = 1; i <= execContext->GetArena()->TotalBytes() / 256; ++i) {
         ptr = ArenaAllocatorMalloc(contextptr, 256);
         EXPECT_NE(ptr, nullptr);
         // check 256 is allocated each call;
